refactor(linux): apply engine settings via funput_configure
Migrate the Fcitx5 and IBus shells to the batch config API, and split ffi_handle.h,
which had grown past the project's 150-LOC-per-file rule (158).

- common/utf8.h (new): decodeUtf8 / appendUtf8 moved out of ffi_handle.h, mirroring
  funput-ffi's abi::codec split. Header-only, resolved via the existing PUBLIC
  include dir, so no CMake change.
- common/ffi_handle.h (158 -> 112): drop the six per-option setters, add
  configure(const Settings&) plus a free engineConfig(const Settings&) that maps the
  settings model to FunputConfig. Both shells now share that one mapping instead of
  restating each option, so a new engine option is added in exactly one place.
- fcitx5 applySettings / ibus applySettings: six setter calls -> handle.configure(settings).

setEnabled stays separate: VI/EN is per-app runtime state (the toggle and the per-app
default drive it), not a persisted preference in the batch config.

Verified on the dev Mac: the shared headers compile with -Wall -Wextra -Werror, and
the Settings -> FunputConfig mapping is checked by a *linked and executed* assertion
program (engineConfig is pure), so a swapped field cannot pass as it could with a
compile-only check. The fcitx5/ibus .cpp files still need a real Linux build.
