+++ platforms/linux/common/ffi_handle.h
@@ -1,6 +1,10 @@
 // RAII C++ wrapper around the `funput-ffi` C ABI (crates/funput-ffi/include/funput.h).
 // Mirrors macOS's FunputComposer.swift: owns one opaque `FunputEngine*`, feeds it
-// UTF-32 scalars, and decodes the composed buffer back to UTF-8 for Fcitx5.
+// UTF-32 scalars, and decodes the composed buffer back to UTF-8 for Fcitx5/IBus.
+//
+// UTF-8 marshalling lives in utf8.h; `configure` takes a whole [Settings] so both
+// shells share one mapping instead of restating every option (add a new engine
+// option in `engineConfig` below and both Fcitx5 and IBus pick it up).
 
 #ifndef FUNPUT_FFI_HANDLE_H
 #define FUNPUT_FFI_HANDLE_H
@@ -10,6 +14,8 @@
 #include <vector>
 
 #include "funput.h" // generated by cbindgen; on the include path via CMake
+#include "settings.h"
+#include "utf8.h"
 
 namespace funput {
 
@@ -21,68 +27,18 @@ enum class KeySource : uint32_t {
     Numpad = SOURCE_NUMPAD,
 };
 
-// Decode a UTF-8 string into Unicode scalars (the inverse of appendUtf8). Invalid
-// bytes are skipped. Used to marshal gõ tắt strings to the FFI's uint32_t arrays.
-inline std::vector<uint32_t> decodeUtf8(const std::string &s) {
-    std::vector<uint32_t> out;
-    size_t i = 0;
-    const size_t n = s.size();
-    while (i < n) {
-        const auto b0 = static_cast<unsigned char>(s[i]);
-        uint32_t cp = 0;
-        size_t extra = 0;
-        if (b0 < 0x80) {
-            cp = b0;
-        } else if ((b0 & 0xE0) == 0xC0) {
-            cp = b0 & 0x1F;
-            extra = 1;
-        } else if ((b0 & 0xF0) == 0xE0) {
-            cp = b0 & 0x0F;
-            extra = 2;
-        } else if ((b0 & 0xF8) == 0xF0) {
-            cp = b0 & 0x07;
-            extra = 3;
-        } else {
-            ++i; // invalid lead byte — skip
-            continue;
-        }
-        if (i + extra >= n) break; // truncated sequence
-        bool ok = true;
-        for (size_t k = 1; k <= extra; ++k) {
-            const auto bk = static_cast<unsigned char>(s[i + k]);
-            if ((bk & 0xC0) != 0x80) {
-                ok = false;
-                break;
-            }
-            cp = (cp << 6) | (bk & 0x3F);
-        }
-        if (ok) {
-            out.push_back(cp);
-            i += extra + 1;
-        } else {
-            ++i; // malformed continuation — skip the lead byte
-        }
-    }
-    return out;
-}
-
-// Append one Unicode scalar to a UTF-8 string.
-inline void appendUtf8(std::string &out, uint32_t cp) {
-    if (cp < 0x80) {
-        out.push_back(static_cast<char>(cp));
-    } else if (cp < 0x800) {
-        out.push_back(static_cast<char>(0xC0 | (cp >> 6)));
-        out.push_back(static_cast<char>(0x80 | (cp & 0x3F)));
-    } else if (cp < 0x10000) {
-        out.push_back(static_cast<char>(0xE0 | (cp >> 12)));
-        out.push_back(static_cast<char>(0x80 | ((cp >> 6) & 0x3F)));
-        out.push_back(static_cast<char>(0x80 | (cp & 0x3F)));
-    } else {
-        out.push_back(static_cast<char>(0xF0 | (cp >> 18)));
-        out.push_back(static_cast<char>(0x80 | ((cp >> 12) & 0x3F)));
-        out.push_back(static_cast<char>(0x80 | ((cp >> 6) & 0x3F)));
-        out.push_back(static_cast<char>(0x80 | (cp & 0x3F)));
-    }
+// The engine-affecting options of `settings`, as the C ABI's by-value config. The
+// runtime VI/EN state is deliberately absent: it is per-app (see `setEnabled`), not
+// a persisted preference.
+inline FunputConfig engineConfig(const Settings &s) {
+    return FunputConfig{
+        static_cast<uint8_t>(s.method),
+        static_cast<uint8_t>(s.toneStyle),
+        s.smartRestore,
+        s.eagerRestore,
+        s.spellCheck,
+        s.autoCapitalize,
+    };
 }
 
 // Owns the engine handle; non-copyable, movable.
@@ -103,13 +59,11 @@ public:
         return *this;
     }
 
-    void setMethod(uint8_t method) { funput_set_method(engine_, method); }
-    void setToneStyle(uint8_t style) { funput_set_tone_style(engine_, style); }
+    // Apply every engine option in one FFI call. Same side effects the individual
+    // setters had: a method change clears the composition, and turning
+    // auto-capitalize off resets its tracking.
+    void configure(const Settings &s) { funput_configure(engine_, engineConfig(s)); }
     void setEnabled(bool on) { funput_set_enabled(engine_, on); }
-    void setSmartRestore(bool on) { funput_set_smart_restore(engine_, on); }
-    void setEagerRestore(bool on) { funput_set_eager_restore(engine_, on); }
-    void setSpellCheck(bool on) { funput_set_spell_check(engine_, on); }
-    void setAutoCapitalize(bool on) { funput_set_auto_capitalize(engine_, on); }
     void armCapitalization() { funput_arm_capitalization(engine_); }
     void clear() { funput_clear(engine_); }
